+++ Scopy/Views/History/HistoryItemView.swift
@@ -254,7 +254,7 @@ struct HistoryItemView: View, Equatable {
                         self.previewModel.previewCGImage = nil
                         self.previewModel.text = nil  // v0.15: Reset text preview content
                         self.previewModel.markdownHTML = nil
-                        self.previewModel.markdownContentHeight = nil
+                        self.previewModel.markdownContentSize = nil
                         self.previewModel.isMarkdown = false
                     }
                 }
@@ -301,7 +301,7 @@ struct HistoryItemView: View, Equatable {
                                 self.showTextPreview = false
                                 self.previewModel.text = nil
                                 self.previewModel.markdownHTML = nil
-                                self.previewModel.markdownContentHeight = nil
+                                self.previewModel.markdownContentSize = nil
                                 self.previewModel.isMarkdown = false
                             }
                         }
@@ -331,7 +331,7 @@ struct HistoryItemView: View, Equatable {
             previewModel.previewCGImage = nil
             previewModel.text = nil
             previewModel.markdownHTML = nil
-            previewModel.markdownContentHeight = nil
+            previewModel.markdownContentSize = nil
             previewModel.isMarkdown = false
         }
         .onChange(of: isScrolling) { _, newValue in
@@ -348,7 +348,7 @@ struct HistoryItemView: View, Equatable {
             previewModel.previewCGImage = nil
             previewModel.text = nil
             previewModel.markdownHTML = nil
-            previewModel.markdownContentHeight = nil
+            previewModel.markdownContentSize = nil
             previewModel.isMarkdown = false
         }
     }
@@ -496,7 +496,7 @@ struct HistoryItemView: View, Equatable {
                 self.previewModel.text = preview
                 self.previewModel.isMarkdown = isMarkdown
                 self.previewModel.markdownHTML = nil
-                self.previewModel.markdownContentHeight = nil
+                self.previewModel.markdownContentSize = nil
                 self.showTextPreview = true
             }